projeto03/requirements.txt
python-dotenv
openai
groq
numpy
PyPDF2
python-docx